+++ src/components/serverManager/serverManager.cpp
@@ -234,16 +234,19 @@ NS_IMETHODIMP ServerManager::IsRunning(mozbool *retVal) {
 NS_IMETHODIMP ServerManager::Stop(mozbool *retVal) {
   *retVal = PR_TRUE;
 
-#ifdef _WIN32
-  HANDLE ps = OpenProcess( SYNCHRONIZE|PROCESS_TERMINATE, 
-			   FALSE, this->serverPid);
-  TerminateProcess(ps, 0);
-#else
-  kill(this->serverPid, SIGTERM);
-  this->url = "";
-#endif
-  this->serverPid = 0;
-  
+if (this->serverPid != 0) {
+    #ifdef _WIN32
+      HANDLE ps = OpenProcess( SYNCHRONIZE|PROCESS_TERMINATE, 
+    			   FALSE, this->serverPid);
+      TerminateProcess(ps, 0);
+    #else
+      kill(this->serverPid, SIGTERM);
+      this->url = "";
+    #endif
+      this->serverPid = 0;
+} else {
+    *retVal = PR_FALSE;
+}  
   return NS_OK;
 }
